+++ src/compositionRoot.ts
@@ -29,6 +29,13 @@ import { GenerateSortOrderUseCase } from "./domain/reports/mal-data-approval/use
 import { GetMonitoringUseCase } from "./domain/reports/mal-data-approval/usecases/GetMonitoringUseCase";
 import { SaveMonitoringUseCase } from "./domain/reports/mal-data-approval/usecases/SaveMonitoringUseCase";
 import { DuplicateDataValuesUseCase } from "./domain/reports/mal-data-approval/usecases/DuplicateDataValuesUseCase";
+import { GetMalSubscriptionDataSetsUseCase } from "./domain/reports/mal-data-subscription/usecases/GetMalSubscriptionDataSetsUseCase";
+import { SaveMalDataSubscriptionColumnsUseCase } from "./domain/reports/mal-data-subscription/usecases/SaveMalDataSubscriptionColumnsUseCase";
+import { GetMalDataSubscriptionColumnsUseCase } from "./domain/reports/mal-data-subscription/usecases/GetMalDataSubscriptionColumnsUseCase";
+import { SaveMalDataSetsSubscriptionUseCase } from "./domain/reports/mal-data-subscription/usecases/SaveMalDataSetsSubscriptionUseCase";
+import { GetSubscriptionSortOrderUseCase } from "./domain/reports/mal-data-subscription/usecases/GetSubscriptionSortOrderUseCase";
+import { GenerateSubscriptionSortOrderUseCase } from "./domain/reports/mal-data-subscription/usecases/GenerateSubscriptionSortOrderUseCase";
+import { UpdateMalSubscriptionStatusUseCase } from "./domain/reports/mal-data-subscription/usecases/UpdateMalSubscriptionStatusUseCase";
 
 export function getCompositionRoot(api: D2Api) {
     const configRepository = new Dhis2ConfigRepository(api, getReportType());
@@ -69,17 +76,13 @@ export function getCompositionRoot(api: D2Api) {
             generateSortOrder: new GenerateSortOrderUseCase(dataDuplicationRepository),
         }),
         malDataSubscription: getExecute({
-            get: new GetMalDataSetsUseCase(dataSubscriptionRepository),
-            getDiff: new GetMalDataDiffUseCase(dataSubscriptionRepository),
-            save: new SaveMalDataSetsUseCase(dataSubscriptionRepository),
-            getColumns: new GetMalDataApprovalColumnsUseCase(dataSubscriptionRepository),
-            saveColumns: new SaveMalDataApprovalColumnsUseCase(dataSubscriptionRepository),
-            getMonitoring: new GetMonitoringUseCase(dataSubscriptionRepository),
-            saveMonitoring: new SaveMonitoringUseCase(dataSubscriptionRepository),
-            updateStatus: new UpdateMalApprovalStatusUseCase(dataSubscriptionRepository),
-            duplicateValue: new DuplicateDataValuesUseCase(dataSubscriptionRepository),
-            getSortOrder: new GetSortOrderUseCase(dataSubscriptionRepository),
-            generateSortOrder: new GenerateSortOrderUseCase(dataSubscriptionRepository),
+            get: new GetMalSubscriptionDataSetsUseCase(dataSubscriptionRepository),
+            save: new SaveMalDataSetsSubscriptionUseCase(dataSubscriptionRepository),
+            getColumns: new GetMalDataSubscriptionColumnsUseCase(dataSubscriptionRepository),
+            saveColumns: new SaveMalDataSubscriptionColumnsUseCase(dataSubscriptionRepository),
+            updateStatus: new UpdateMalSubscriptionStatusUseCase(dataSubscriptionRepository),
+            getSortOrder: new GetSubscriptionSortOrderUseCase(dataSubscriptionRepository),
+            generateSortOrder: new GenerateSubscriptionSortOrderUseCase(dataSubscriptionRepository),
         }),
         orgUnits: getExecute({
             get: new GetOrgUnitsUseCase(orgUnitsRepository),
