--- src/data/reports/mal-data-subscription/sql-views/mal-data-approval-header.sql
@@ -1,9 +1,0 @@
-SELECT dataset.name                                          AS dataset,
-       dataset.uid                                                       AS datasetuid,
-       organisationunit.uid                                       AS orgunituid,
-       organisationunit.name                                  AS orgunit
-       FROM dataset
-                JOIN datasetsource USING (datasetid)
-                JOIN organisationunit ON (organisationunitid=sourceid)
-where dataset.uid    ~ ('^' || replace('${dataSets}', '-', '|') || '$')
-group by dataset.uid, dataset.name, organisationunit.name, organisationunit.uid
