+++ src/domain/reports/mal-data-subscription/usecases/GenerateSubscriptionSortOrderUseCase.ts
@@ -1,10 +1,10 @@
 import { UseCase } from "../../../../compositionRoot";
 import { MalDataSubscriptionRepository } from "../repositories/MalDataSubscriptionRepository";
 
-export class GetSortOrderUseCase implements UseCase {
+export class GenerateSubscriptionSortOrderUseCase implements UseCase {
     constructor(private approvalRepository: MalDataSubscriptionRepository) {}
 
-    execute(): Promise<string[]> {
-        return this.approvalRepository.getSortOrder();
+    execute(): Promise<void> {
+        return this.approvalRepository.generateSortOrder();
     }
 }
