+++ src/domain/reports/mal-data-subscription/usecases/GetMalDataSubscriptionColumnsUseCase.ts
@@ -1,7 +1,7 @@
 import { UseCase } from "../../../../compositionRoot";
 import { MalDataSubscriptionRepository } from "../repositories/MalDataSubscriptionRepository";
 
-export class GetMalDataApprovalColumnsUseCase implements UseCase {
+export class GetMalDataSubscriptionColumnsUseCase implements UseCase {
     constructor(private approvalRepository: MalDataSubscriptionRepository) {}
 
     execute(namespace: string): Promise<string[]> {
