+++ src/domain/reports/mal-data-subscription/usecases/GetSubscriptionSortOrderUseCase.ts
@@ -1,10 +1,10 @@
 import { UseCase } from "../../../../compositionRoot";
 import { MalDataSubscriptionRepository } from "../repositories/MalDataSubscriptionRepository";
 
-export class GenerateSortOrderUseCase implements UseCase {
+export class GetSubscriptionSortOrderUseCase implements UseCase {
     constructor(private approvalRepository: MalDataSubscriptionRepository) {}
 
-    execute(): Promise<void> {
-        return this.approvalRepository.generateSortOrder();
+    execute(): Promise<string[]> {
+        return this.approvalRepository.getSortOrder();
     }
 }
