+++ src/domain/reports/mal-data-subscription/usecases/SaveMalDataSetsSubscriptionUseCase.ts
@@ -2,7 +2,7 @@ import { UseCase } from "../../../../compositionRoot";
 import { MalDataSubscriptionItem } from "../entities/MalDataSubscriptionItem";
 import { MalDataSubscriptionRepository } from "../repositories/MalDataSubscriptionRepository";
 
-export class SaveMalDataSetsUseCase implements UseCase {
+export class SaveMalDataSetsSubscriptionUseCase implements UseCase {
     constructor(private dataSetRepository: MalDataSubscriptionRepository) {}
 
     async execute(filename: string, dataSets: MalDataSubscriptionItem[]): Promise<void> {
