+++ src/domain/reports/mal-data-subscription/usecases/SaveMalDataSubscriptionColumnsUseCase.ts
@@ -1,7 +1,7 @@
 import { UseCase } from "../../../../compositionRoot";
 import { MalDataSubscriptionRepository } from "../repositories/MalDataSubscriptionRepository";
 
-export class SaveMalDataApprovalColumnsUseCase implements UseCase {
+export class SaveMalDataSubscriptionColumnsUseCase implements UseCase {
     constructor(private approvalRepository: MalDataSubscriptionRepository) {}
 
     execute(namespace: string, columns: string[]): Promise<void> {
