+++ src/webapp/reports/Reports.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { AdminReport } from "./admin/AdminReport";
 import MalDataApprovalStatusReport from "./mal-data-approval/MalDataApprovalReport";
-import MalDataSubscriptionStatusReport from "./mal-data-subscription/MalDataSubmissionReport";
+import MalDataSubscriptionStatusReport from "./mal-data-subscription/MalDataSubscriptionReport";
 import { NHWADataApprovalStatusReport } from "./nhwa-approval-status/NHWADataApprovalStatusReport";
 import { NHWACommentsReport } from "./nhwa-comments/NHWACommentsReport";
 import { WMRNationalPolicies } from "./wmr-national-policies/WMRNationalPolicies";
