+++ src/webapp/reports/mal-data-subscription/DataSubscriptionViewModel.ts
@@ -1,9 +1,8 @@
 import { Config } from "../../../domain/common/entities/Config";
 import {
     MalDataSubscriptionItem,
-    getDataDuplicationItemId,
+    getDataSubscriptionItemId,
 } from "../../../domain/reports/mal-data-subscription/entities/MalDataSubscriptionItem";
-import { toDate } from "date-fns-tz";
 
 export interface DataSubscriptionViewModel {
     id: string;
@@ -12,15 +11,6 @@ export interface DataSubscriptionViewModel {
     orgUnitUid: string;
     orgUnit: string;
     period: string;
-    attribute: string;
-    approvalWorkflowUid: string;
-    approvalWorkflow: string;
-    completed: boolean;
-    validated: boolean;
-    lastUpdatedValue: Date | undefined;
-    lastDateOfSubmission: Date | undefined;
-    lastDateOfApproval: Date | undefined;
-    modificationCount: string | undefined;
 }
 
 export function getDataSubscriptionViews(
@@ -29,25 +19,12 @@ export function getDataSubscriptionViews(
 ): DataSubscriptionViewModel[] {
     return items.map(item => {
         return {
-            id: getDataDuplicationItemId(item),
+            id: getDataSubscriptionItemId(item),
             dataSetUid: item.dataSetUid,
             dataSet: item.dataSet,
             orgUnitUid: item.orgUnitUid,
             orgUnit: item.orgUnit,
             period: item.period,
-            attribute: item.attribute ?? "-",
-            approvalWorkflowUid: item.approvalWorkflowUid ?? "-",
-            approvalWorkflow: item.approvalWorkflow ?? "-",
-            completed: item.completed,
-            validated: item.validated,
-            lastUpdatedValue: item.lastUpdatedValue ? toDate(item.lastUpdatedValue, { timeZone: "UTC" }) : undefined,
-            lastDateOfSubmission: item.lastDateOfSubmission
-                ? toDate(item.lastDateOfSubmission, { timeZone: "UTC" })
-                : undefined,
-            lastDateOfApproval: item.lastDateOfApproval
-                ? toDate(item.lastDateOfApproval, { timeZone: "UTC" })
-                : undefined,
-            modificationCount: item.modificationCount,
         };
     });
 }
