+++ src/webapp/reports/mal-data-subscription/MalDataSubscriptionReport.tsx
@@ -2,7 +2,7 @@ import { Typography, makeStyles } from "@material-ui/core";
 import { useEffect } from "react";
 import i18n from "../../../locales";
 import { useAppContext } from "../../contexts/app-context";
-import { DataSubscriptionList } from "./data-approval-list/DataSubscriptionList";
+import { DataSubscriptionList } from "./data-subscription-list/DataSubscriptionList";
 
 const MalDataSubscriptionStatusReport: React.FC = () => {
     const classes = useStyles();
